W1 HDQ: Modify read call to support reporting errors
When no battery is present for the BQ27000, attempting to
read registers causes a timeout on the bus.  The BQ27000
driver supports reporting "POWER_SUPPLY_PRESENT=0", but
this only works if the read/write calls return an error
when they couldn't communicate.  By providing this through
the W1 layer, the driver in drivers/power/bq27x00_battery.c
can properly report if a battery is present or not on
the system.

git-svn-id: svn+ssh://svn/cvs/eps_svn/software/products/linux/kernels/linux-3.0@27953 d7e73fac-2431-0410-abd5-fd325483733a

+++ drivers/w1/slaves/w1_bq27000.c
@@ -28,7 +28,7 @@ static int F_ID;
 
 static int w1_bq27000_read(struct device *dev, unsigned int reg)
 {
-	u8 val;
+	int val;
 	struct w1_slave *sl = container_of(dev->parent, struct w1_slave, dev);
 
 	mutex_lock(&sl->master->mutex);
